rl-playground/pyproject.toml
[project]
name = "rl-playground" 
version = "0.1.0" 
description = "Add your description here" 
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "gradio>=5.44.1",
    "gymnasium[classic-control]>=1.2.0",
    "imageio[ffmpeg]>=2.37.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "stable-baselines3[extra]>=2.7.0",
    "tensorboard>=2.20.0",
    "tqdm>=4.67.1",
    "ufal.pybox2d>=2.3.10.4",
    "torch==2.5.1+cu121"
]


[[tool.uv.index]]
name = "pypi"
url  = "https://pypi.org/simple"

[[tool.uv.index]]
name = "pytorch-cu121"
url  = "https://download.pytorch.org/whl/cu121"

[tool.uv.sources]
torch       = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }
torchaudio  = { index = "pytorch-cu121" }




